tehnezis_bot/pyproject.toml
[tool.poetry]
name = "tehnezis-bot"  # Или другое имя вашего проекта
version = "0.1.0"
description = "bot для обработки Excel" 
authors = ["gob26 profservis26@yandex.ru>"] 

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyTelegramBotAPI = "^4.24.0"
pandas = "^2.2.3"
openpyxl = "^3.1.3"
lxml = "^5.3.0"
aiohttp = "^3.11.16"
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
python-dotenv = "^1.1.0"
alembic = "^1.15.2"
docker = "^7.1.0"
aiogram = "^3.20.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"